Update README.md
This set of programming exercises focuses on implementing various functions to solve mathematical problems. It includes tasks such as identifying even and prime numbers, calculating factorials, determining multiples, and computing binomial coefficients. Additionally, it covers the implementation of a series expansion to approximate the exponential function. The exercises are designed to enhance understanding of function development, loops, and mathematical algorithms in programming